+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 - Corected equality checks.
 - Only numbers and strings can be concatenated.
 - Wrote a bunch of test scripts to begin building a test suite.
+- Changed `showDateTime` to `dateTimeFormat` which is a `String?` type.
+  - When this is null, it does not show the date.
+  - When this is non-null, uses intl package `DateFormat` class.
+    - See: https://pub.dev/documentation/intl/latest/intl/DateFormat-class.html
+- Added `showSidebarIndex` which now paritions the page so that the index is in a sticky sidebar.
+- Title, version, and datetime now have css classes `version-title`, `version-number`, `version-datetime` respectively.
+- 
 
 ## 1.0.2
 - Fixed variables without parent objects (non-fields + globals) not generating headers.